refactor leader into multiple files
